// src/model/user/IUser.ts
import {Hair} from "./IHair.ts";
import {Address} from "./IAddress.ts";
import {Bank} from "./IBank.ts";
import {Company} from "./ICompany.ts";
import {Crypto} from "./ICrypto.ts";

export interface IUser {
  id: number;
  firstName: string;
  lastName: string;
  maidenName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: Hair;
  ip: string;
  address: Address;
  macAddress: string;
  university: string;
  bank: Bank;
  company: Company;
  ein: string;
  ssn: string;
  userAgent: string;
  crypto: Crypto;
  role: string;
}